Deep Learning_ITAI2376/requirements.txt
python-dotenv==1.2.1
azure-core==1.39.0
azure-ai-documentintelligence==1.0.2
langchain==1.2.0
langchain-core==1.3.0
langchain-openai==1.1.6
langchain-community==0.4.1
langchain-text-splitters==1.1.2
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
rank-bm25==0.2.2
matplotlib==3.10.8
